+++ subsystems/memory/heap.h
@@ -5,7 +5,7 @@ class Heap {
 public:
 	virtual void init(long size); // size in number of ints (4 bytes)
 	virtual void free(void* ptr);
-	virtual void* malloc(size_t size);
+	virtual void* malloc(int size);
 };
 
 #endif
